Added Backend connection for requests
Former-commit-id: 058a1caa83d5ec35fb020ec5b4402696c63867d8

+++ carefuel_extended/src/main/java/carefuel/path/Graph.java
@@ -23,7 +23,7 @@ public class Graph<E> {
 
 	private List<E> values;
 	private List<Vertex<E>> vertices;
-	private Float[][] distances;
+	private float[][] distances;
 	private int size;
 
 	/**
@@ -33,7 +33,7 @@ public class Graph<E> {
 	 * @param distances
 	 * @throws Exception
 	 */
-	public Graph(List<E> values, Float[][] distances) throws Exception {
+	public Graph(List<E> values, float[][] distances) throws Exception {
 		if (values.size() != distances.length) {
 			throw new Exception("Dimensions of vertices and distances don't get along!");
 		}
@@ -47,7 +47,7 @@ public class Graph<E> {
 	 * 
 	 * @param distances
 	 */
-	private void setDistances(Float[][] distances) {
+	private void setDistances(float[][] distances) {
 		if (distances.length > 0 && distances.length == size) {
 			if (distances.length != distances[0].length) {
 				log.error("Dimension mismatch in distances: Shape (" + distances.length + "," + distances[0].length
@@ -70,7 +70,7 @@ public class Graph<E> {
 		return this.vertices;
 	}
 
-	public Float[][] getDistances() {
+	public float[][] getDistances() {
 		return distances;
 	}
 
@@ -81,7 +81,7 @@ public class Graph<E> {
 	 * @param maxRange
 	 * @return
 	 */
-	public PriorityQueue<Edge<E>> getNeighbours(Vertex<E> node, short maxRange) {
+	public PriorityQueue<Edge<E>> getNeighbours(Vertex<E> node, float maxRange) {
 
 		PriorityQueue<Edge<E>> neighbours = node.getNeighbours();
 		// Try to use already computed neighbours
@@ -96,7 +96,7 @@ public class Graph<E> {
 		}
 
 		// distances to all neighbours
-		Float[] neighbourDistances = distances[values.indexOf(node.getValue())];
+		float[] neighbourDistances = distances[values.indexOf(node.getValue())];
 
 		// for all vertices...
 		for (int i = 0; i < neighbourDistances.length; i++) {
